Flyttet admin over til sin egen fil
